001-basic: remove TaskEither example
It's confusing for fp-ts devs because they then start thinking about
TaskEither and ReaderTaskEither, and try to extract a 1:1 mapping of the
Effect type with their current mental model.

OTOH it's immediately apparent that an Effect is not really a 1:1 mapping to
the provided type, but it's just a signature to drive the point across.